Moving AtB Weight Generation into a static class instead of calling the unit market
