+++ MekHQ/src/mekhq/campaign/mission/atb/scenario/AllyRescueBuiltInScenario.java
@@ -23,7 +23,7 @@ import megamek.common.Entity;
 import megamek.common.EntityWeightClass;
 import megamek.common.UnitType;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.market.unitMarket.AtBMonthlyUnitMarket;
+import mekhq.campaign.againstTheBot.AtBStaticWeightGenerator;
 import mekhq.campaign.mission.*;
 import mekhq.campaign.mission.ObjectiveEffect.ObjectiveEffectType;
 import mekhq.campaign.mission.ScenarioObjective.ObjectiveCriterion;
@@ -92,7 +92,7 @@ public class AllyRescueBuiltInScenario extends AtBScenario {
         for (int i = 0; i < 4; i++) {
             getAlliesPlayer().add(getEntity(contract.getEmployerCode(), contract.getAllySkill(),
                     contract.getAllyQuality(), UnitType.MEK,
-                    AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, contract.getEmployerFaction()),
+                    AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, contract.getEmployerFaction()),
                     campaign));
         }
 
@@ -101,7 +101,7 @@ public class AllyRescueBuiltInScenario extends AtBScenario {
         for (int i = 0; i < 8; i++) {
             int weightClass;
             do {
-                weightClass = AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, contract.getEmployerFaction());
+                weightClass = AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, contract.getEmployerFaction());
             } while (weightClass >= EntityWeightClass.WEIGHT_ASSAULT);
             otherForce.add(getEntity(contract.getEmployerCode(), contract.getAllySkill(),
                     contract.getAllyQuality(), UnitType.MEK, weightClass, campaign));
@@ -112,7 +112,7 @@ public class AllyRescueBuiltInScenario extends AtBScenario {
         for (int i = 0; i < 12; i++) {
             int weightClass;
             do {
-                weightClass = AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, contract.getEnemy());
+                weightClass = AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, contract.getEnemy());
             } while (weightClass <= EntityWeightClass.WEIGHT_LIGHT);
             enemyEntities.add(getEntity(contract.getEnemyCode(), contract.getEnemySkill(),
                     contract.getEnemyQuality(), UnitType.MEK, weightClass, campaign));
