+++ MekHQ/src/mekhq/campaign/mission/atb/scenario/PirateFreeForAllBuiltInScenario.java
@@ -18,25 +18,21 @@
  */
 package mekhq.campaign.mission.atb.scenario;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import megamek.common.Board;
 import megamek.common.Entity;
 import megamek.common.EntityWeightClass;
 import megamek.common.UnitType;
 import megamek.common.enums.SkillLevel;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.market.unitMarket.AtBMonthlyUnitMarket;
-import mekhq.campaign.mission.AtBContract;
-import mekhq.campaign.mission.AtBScenario;
-import mekhq.campaign.mission.BotForce;
-import mekhq.campaign.mission.CommonObjectiveFactory;
-import mekhq.campaign.mission.ScenarioObjective;
+import mekhq.campaign.againstTheBot.AtBStaticWeightGenerator;
+import mekhq.campaign.mission.*;
 import mekhq.campaign.mission.atb.AtBScenarioEnabled;
 import mekhq.campaign.rating.IUnitRating;
 import mekhq.campaign.universe.Faction;
 import mekhq.campaign.universe.Factions;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @AtBScenarioEnabled
 public class PirateFreeForAllBuiltInScenario extends AtBScenario {
@@ -87,7 +83,7 @@ public class PirateFreeForAllBuiltInScenario extends AtBScenario {
         for (int i = 0; i < 4; i++) {
             int weightClass;
             do {
-                weightClass = AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, contract.getEmployerFaction());
+                weightClass = AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, contract.getEmployerFaction());
             } while (weightClass >= EntityWeightClass.WEIGHT_ASSAULT);
             getAlliesPlayer().add(getEntity(contract.getEmployerCode(), contract.getAllySkill(),
                     contract.getAllyQuality(), UnitType.MEK, weightClass, campaign));
@@ -96,7 +92,7 @@ public class PirateFreeForAllBuiltInScenario extends AtBScenario {
         for (int i = 0; i < 12; i++) {
             enemyEntities.add(getEntity(contract.getEnemyCode(), contract.getEnemySkill(),
                     contract.getEnemyQuality(), UnitType.MEK,
-                    AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, contract.getEnemy()),
+                    AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, contract.getEnemy()),
                     campaign));
         }
 
@@ -107,7 +103,7 @@ public class PirateFreeForAllBuiltInScenario extends AtBScenario {
         for (int i = 0; i < 12; i++) {
             otherForce.add(getEntity(faction.getShortName(), SkillLevel.REGULAR,
                     IUnitRating.DRAGOON_C, UnitType.MEK,
-                    AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, faction), campaign));
+                    AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, faction), campaign));
         }
 
         addBotForce(new BotForce(PIRATE_FORCE_ID, 3, Board.START_S, otherForce), campaign);
