+++ MekHQ/src/mekhq/campaign/mission/atb/scenario/StarLeagueCache1BuiltInScenario.java
@@ -19,22 +19,11 @@
 package mekhq.campaign.mission.atb.scenario;
 
 import megamek.client.generator.RandomUnitGenerator;
-import megamek.common.Board;
-import megamek.common.Compute;
-import megamek.common.Entity;
-import megamek.common.EntityWeightClass;
-import megamek.common.MechSummary;
-import megamek.common.UnitType;
+import megamek.common.*;
 import megamek.common.enums.SkillLevel;
 import mekhq.campaign.Campaign;
-import mekhq.campaign.market.unitMarket.AtBMonthlyUnitMarket;
-import mekhq.campaign.mission.AtBContract;
-import mekhq.campaign.mission.AtBDynamicScenarioFactory;
-import mekhq.campaign.mission.AtBScenario;
-import mekhq.campaign.mission.BotForce;
-import mekhq.campaign.mission.CommonObjectiveFactory;
-import mekhq.campaign.mission.Loot;
-import mekhq.campaign.mission.ScenarioObjective;
+import mekhq.campaign.againstTheBot.AtBStaticWeightGenerator;
+import mekhq.campaign.mission.*;
 import mekhq.campaign.mission.atb.AtBScenarioEnabled;
 import mekhq.campaign.rating.IUnitRating;
 import mekhq.campaign.universe.Faction;
@@ -129,7 +118,7 @@ public class StarLeagueCache1BuiltInScenario extends AtBScenario {
             // TODO : AtB Star League RAT Roll Year Option
             final Faction faction = Factions.getInstance().getFaction("SL");
             ms = campaign.getUnitGenerator().generate(faction.getShortName(), UnitType.MEK,
-                    AtBMonthlyUnitMarket.getRandomWeight(campaign, UnitType.MEK, faction), 2750,
+                    AtBStaticWeightGenerator.getRandomWeight(campaign, UnitType.MEK, faction), 2750,
                     (roll == 6) ? IUnitRating.DRAGOON_A : IUnitRating.DRAGOON_D);
         }
         Entity en = (ms == null) ? null
